# api/openapi/auth-service.openapi.yaml
openapi: 3.1.0
info:
  title: Vista Platform — auth-service API
  version: "0.1.0"
  description: |
    Spec-first contract for auth-service (ADR-0001). Eighth (and final
    pre-spec) vertical slice for the queue: covers the three **cross-cutter**
    endpoints every page in any future surface calls on load.

    ## Scope of this slice

    Three endpoints on the v1 path prefix `/api/v1/auth-service`:

      - `GET /auth/me`           — current user + tenant context (web-ui `auth-context.tsx` on every login + bootstrap)
      - `GET /user/permissions`  — string-array of effective permissions for RBAC gating (web-ui `PermissionGate` / `usePermissions()`)
      - `GET /tenant/features`   — resolved feature-flag map plus usage limits (web-ui `useFeatures()` / `useFeature()`)

    Out of scope for this slice (each is a larger surface that needs its own
    slice with the design track's input):

      - Login / logout / register / refresh / forgot-password / reset-password
      - SSO authorize / callback / link / unlink (both tenant and platform)
      - `/auth/me` writes (PUT, change-password), preferences, EULA, onboarding
      - Sessions, connections, avatar uploads
      - Users-admin CRUD, RBAC role + permission management, permission-check POST
      - Billing, tier selection, subscription tier reads
      - Tenant read/update, branding, UI config
      - Platform-admin and impersonation surfaces

...
    fetches on load). The intro above is kept as the original framing.

...

...

...
servers:
...
paths:
...
  /user/permissions:
    get:
      operationId: getCurrentUserPermissions
      tags: [permissions]
      summary: Get the current user's effective permission names
      description: >
        Returns a flat string array of permission names (e.g. `compliance.manage`,
        `users.read`) that the authenticated user has been granted via their
        role assignments under the current tenant. Used by the web-ui
        PermissionGate / PermissionRoute components.
      responses:
        "200":
          description: The user's effective permission names.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionsResponse"
        "400":
          $ref: "#/components/responses/LegacyBadRequest"
        "401":
          $ref: "#/components/responses/LegacyUnauthorized"
        "500":
          $ref: "#/components/responses/LegacyServerError"
      x-quirks:
        - "200 with empty permissions: when a user has no role assignments (e.g. a fresh tenant or a not-yet-provisioned admin), the handler returns `{ \"permissions\": [] }`. The web-ui's hasPermission()/hasAnyPermission()/hasAllPermissions() helpers all treat the empty array as the safe (deny-all) default."
        - "Permissions are returned as plain strings (not the rich Permission objects with id/description/resource/action). For the resource/action breakdown, callers must use a different endpoint (e.g. `/permissions` with tenant-admin role)."
...
components:
...
  responses:
    LegacyBadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
    LegacyUnauthorized:
      description: Missing or invalid auth / tenant context.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
...
    LegacyServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
  schemas:
...
    PermissionsResponse:
      type: object
      additionalProperties: false
      description: >
        CURRENT response shape for `GET /user/permissions`. A flat string
        array under the `permissions` key. Hardening to ADR-0002 would wrap
        as `{ data: [...], pagination: {...} }`.
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            type: string
          description: Permission names (e.g. `compliance.manage`, `users.read`).

    # ============================================================
    # /tenant/features
    # ============================================================
...
    LegacyError:
      type: object
      additionalProperties: true
      description: >
        CURRENT error shape returned by these endpoints today: a single
        human-readable string under `error`. Some 400s carry extra sibling keys
        (e.g. `details`, `required_permission`), so additionalProperties is
        permitted here. Unstructured — cannot be branched on safely. Superseded
        by the ADR-0002 `Error` envelope below as endpoints are hardened.
      required: [error]
      properties:
        error:
          type: string
...